Bail out if node is already installed

+++ Makefile
@@ -1,4 +1,5 @@
 install:
+	if hash node > /dev/null 2>&1; then echo "Seems like node.js is alread installed."; exit 1;  fi
 	wget "http://nodejs.org/dist/v0.10.24/node-v0.10.24-linux-arm-pi.tar.gz" -O node.tar.gz && tar xvzf node.tar.gz && rm node.tar.gz
 	sudo mkdir /opt/node
 	sudo mv node*/* /opt/node
